Next step tools moved back to core
