+++ sgr_deep_research/core/next_step_tools.py
